Refactored much of the badly organized code
