+++ src/modules/game/swipe-tags.tsx
