--- src/modules/header/header.tsx
@@ -1,7 +1,0 @@
-export const Header = () => {
-  return (
-    <header className=''>
-
-    </header>
-  )
-}
