--- src/pages/main.page.tsx
@@ -1,10 +1,0 @@
-import Layout from '@/components/layout'
-import HeroSection from '@/modules/hero-secion/hero.section'
-
-export const MainPage = () => {
-  return (
-    <Layout>
-      <HeroSection/>
-    </Layout>
-  )
-}
