+++ src/pages/swipe.page.tsx
